@@ -7,42 +7,81 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use App\Repository\EventDetailsRepository;
+use Symfony\Component\Validator\Constraints as Assert;
 
 #[ORM\Entity(repositoryClass: EventDetailsRepository::class)]
 class EventDetails
 {
-
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column]
     private ?int $id;
-
+    /**
+     * @Assert\NotBlank(message="This field cannot be empty")
+     * @Assert\Length(
+     *      min = 4,
+     *      minMessage = "The name must be at least {{ limit }} characters long"
+     * )
+     */
     #[ORM\Column]
     private ?string $name;
 
-    #[ORM\Column]
+    /**
+     * @Assert\NotBlank(message="This field cannot be empty")
+     * @Assert\Length(
+     *      min = 4,
+     *      minMessage = "The type must be at least {{ limit }} characters long"
+     * )
+     * @Assert\Regex(
+     *     pattern="/^[a-zA-Z]*$/",
+     *     message="The type should contain only letters"
+     * )
+     */
+    #[ORM\Column(nullable: false)]
     private ?string $type;
 
-    #[ORM\Column]
+    /**
+     * @Assert\NotBlank(message="This field cannot be empty")
+     * @Assert\GreaterThan(
+     *     "now",
+     *     message="The event date should be at least 1 hour from now"
+     * )
+     */
+    #[ORM\Column(nullable: false)]
     private ?\DateTime $eventDate;
 
-    #[ORM\Column]
+    /**
+     * @Assert\NotBlank(message="This field cannot be empty")
+     * @Assert\Range(
+     *      min = 20,
+     *      max = 120,
+     *      notInRangeMessage = "The duration must be between {{ min }} and {{ max }}",
+     * )
+     */
+    #[ORM\Column(nullable: false)]
     private ?string $duree;
 
-    #[ORM\Column]
+    /**
+     * @Assert\NotBlank(message="This field cannot be empty")
+     * @Assert\Range(
+     *      min = 10,
+     *      max = 50,
+     *      notInRangeMessage = "The number of places must be between {{ min }} and {{ max }}",
+     * )
+     */
+    #[ORM\Column(nullable: false)]
     private ?int $nbPlaces;
 
-    #[ORM\Column]
+    #[ORM\Column(nullable: false)]
     private ?int $nbTotal;
 
-    #[ORM\ManyToMany(targetEntity: User::class)]
-    private Collection $user;
-
-
     public function __construct()
-    {
-        $this->user = new \Doctrine\Common\Collections\ArrayCollection();
+    {   
+        $this->user = new ArrayCollection();
     }
+
+    
+
 
     public function getId(): ?int
     {
