@@ -8,69 +8,73 @@ use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use App\Repository\UserRepository;
 use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 
 #[ORM\Entity(repositoryClass: UserRepository::class)]
-class User
+class User Implements UserInterface, PasswordAuthenticatedUserInterface
 {
     #[ORM\Id]
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your CIN.')]
-    #[Assert\Length(min: 8, max: 8, minMessage: 'CIN must be 8 characters long.', maxMessage: 'CIN must be 8 characters long.')]
-    #[Assert\Regex(pattern: '/^[0-9]*$/', message: 'CIN must contain only numbers.')]
+    #[Assert\NotBlank(message: 'Please enter your CIN.', groups: ['create'])]
+    #[Assert\Length(min: 8, max: 8, minMessage: 'CIN must be 8 characters long.', maxMessage: 'CIN must be 8 characters long.', groups: ['create'])]
+    #[Assert\Regex(pattern: '/^[0-9]*$/', message: 'CIN must contain only numbers.', groups: ['create'])]
     private ?int $id;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your username.')]
-    #[Assert\Length(min: 3, max: 20, minMessage: 'Username must be at least 3 characters long.', maxMessage: 'Username must be at most 20 characters long.')]
-    #[Assert\Regex(pattern: '/^[a-zA-Z0-9]*$/', message: 'Username must contain only letters and numbers.')]
+    #[Assert\NotBlank(message: 'Please enter your username.', groups: ['create', 'update'])]
+    #[Assert\Length(min: 3, max: 20, minMessage: 'Username must be at least 3 characters long.', maxMessage: 'Username must be at most 20 characters long.', groups: ['create', 'update'])]
+    #[Assert\Regex(pattern: '/^[a-zA-Z0-9]*$/', message: 'Username must contain only letters and numbers.', groups: ['create', 'update'])]
     private ?string $username;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your firstname.')]
-    #[Assert\Length(min: 3, max: 20, minMessage: 'Firstname must be at least 3 characters long.', maxMessage: 'Firstname must be at most 20 characters long.')]
-    #[Assert\Regex(pattern: '/^[a-zA-Z]*$/', message: 'Firstname must contain only letters.')]
+    #[Assert\NotBlank(message: 'Please enter your firstname.', groups: ['create', 'update'])]
+    #[Assert\Length(min: 3, max: 20, minMessage: 'Firstname must be at least 3 characters long.', maxMessage: 'Firstname must be at most 20 characters long.', groups: ['create', 'update'])]
+    #[Assert\Regex(pattern: '/^[a-zA-Z ]*$/', message: 'Firstname must contain only letters.', groups: ['create', 'update'])]
     private ?string $firstname;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your lastname.')]
-    #[Assert\Length(min: 3, max: 20, minMessage: 'Lastname must be at least 3 characters long.', maxMessage: 'Lastname must be at most 20 characters long.')]
-    #[Assert\Regex(pattern: '/^[a-zA-Z]*$/', message: 'Lastname must contain only letters.')]
+    #[Assert\NotBlank(message: 'Please enter your lastname.', groups: ['create', 'update'])]
+    #[Assert\Length(min: 3, max: 20, minMessage: 'Lastname must be at least 3 characters long.', maxMessage: 'Lastname must be at most 20 characters long.', groups: ['create', 'update'])]
+    #[Assert\Regex(pattern: '/^[a-zA-Z ]*$/', message: 'Lastname must contain only letters.', groups: ['create', 'update'])]
     private ?string $lastname;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your date of birth.')]
-    #[Assert\LessThan('today', message: 'Date of birth must be in the past.')]
-    #[Assert\GreaterThanOrEqual('-100 years', message: 'Date of birth must be at most 100 years ago.')]
+    #[Assert\NotBlank(message: 'Please enter your date of birth.', groups: ['create', 'update'])]
+    #[Assert\LessThan('today', message: 'Date of birth must be in the past.', groups: ['create', 'update'])]
+    #[Assert\GreaterThanOrEqual('-100 years', message: 'Date of birth must be at most 100 years ago.', groups: ['create', 'update'])]
     private ?\DateTime $dateNaiss;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your password.')]
+    #[Assert\NotBlank(message: 'Please enter your password.', groups: ['create', 'login'])]
     #[Assert\Length(min: 8, max: 20, minMessage: 'Password must be at least 8 characters long.', maxMessage: 'Password must be at most 20 characters long.', groups: ['create'])]
     private ?string $password;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your email.')]
-    #[Assert\Email(message: 'Please enter a valid email address.')]
-    #[Assert\Length(min: 3, max: 50, minMessage: 'Email must be at least 3 characters long.', maxMessage: 'Email must be at most 50 characters long.')]
+    #[Assert\NotBlank(message: 'Please enter your email.', groups: ['create', 'login'])]
+    #[Assert\Email(message: 'Please enter a valid email address.', groups: ['create', 'login'])]
+    #[Assert\Length(min: 3, max: 50, minMessage: 'Email must be at least 3 characters long.', maxMessage: 'Email must be at most 50 characters long.', groups: ['create', 'login'])]
     private ?string $email;
 
     #[ORM\Column]
+    #[Assert\NotBlank(message: 'Please enter your role.', groups: ['admincreate'])]
+    #[Assert\Choice(choices: ['admin', 'client', 'staff'], message: 'Role must be either admin or user.', groups: ['admincreate'])]
     private ?string $role;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your phone number.')]
-    #[Assert\Length(min: 8, max: 8, minMessage: 'Phone number must be 8 characters long.', maxMessage: 'Phone number must be 8 characters long.')]
-    #[Assert\Regex(pattern: '/^[0-9]*$/', message: 'Phone number must contain only numbers.')]
+    #[Assert\NotBlank(message: 'Please enter your phone number.', groups: ['create', 'update'])]
+    #[Assert\Length(min: 8, max: 8, minMessage: 'Phone number must be 8 characters long.', maxMessage: 'Phone number must be 8 characters long.', groups: ['create', 'update'])]
+    #[Assert\Regex(pattern: '/^[0-9]*$/', message: 'Phone number must contain only numbers.', groups: ['create', 'update'])]
     private ?string $numTel;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please enter your address.')]
-    #[Assert\Length(min: 3, max: 50, minMessage: 'Address must be at least 3 characters long.', maxMessage: 'Address must be at most 50 characters long.')]
-    #[Assert\Regex(pattern: '/^[a-zA-Z0-9]*$/', message: 'Address must contain only letters and numbers.')]
+    #[Assert\NotBlank(message: 'Please enter your address.', groups: ['create', 'update'])]
+    #[Assert\Length(min: 3, max: 50, minMessage: 'Address must be at least 3 characters long.', maxMessage: 'Address must be at most 50 characters long.', groups: ['create', 'update'])]
+    #[Assert\Regex(pattern: '/^[a-zA-Z0-9 ]*$/', message: 'Address must contain only letters and numbers.', groups: ['create', 'update'])]
     private ?string $adresse;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Please upload your photo.')]
+    #[Assert\NotBlank(message: 'Please upload your photo.', groups: ['create'])]
     #[Assert\File(mimeTypes: ['image/jpeg', 'image/png'], mimeTypesMessage: 'Please upload a valid image file.', groups: ['create'])]
     #[Assert\Image(maxSize: '2M', maxSizeMessage: 'Please upload an image file that is less than 2MB.', groups: ['create'])]
     private ?string $photo;
@@ -84,14 +88,23 @@ class User
     #[ORM\Column]
     private ?\DateTime $faceidTs;
 
-
-    #[ORM\ManyToMany(targetEntity: EventDetails::class, inversedBy: "user")]
-    private Collection $eventDetails;
-
     
     public function __construct()
     {
         $this->eventDetails = new \Doctrine\Common\Collections\ArrayCollection();
+    }
+
+    public function getUserIdentifier(): string
+    {
+        return (string) $this->email;
+    }
+
+    public function getRoles(): array
+    {
+
+        $roles = ["ROLE_".strtoupper($this->role)];
+
+        return array_unique($roles);
     }
 
     public function setId(int $id): static
@@ -153,9 +166,21 @@ class User
         return $this;
     }
 
+
     public function getPassword(): ?string
     {
         return $this->password;
+    }
+
+    public function eraseCredentials(): void
+    {
+        // If you store any temporary, sensitive data on the user, clear it here
+        // $this->plainPassword = null;
+    }
+
+    public function getSalt(): ?string
+    {
+        return null;
     }
 
     public function setPassword(?string $password): static
