@@ -22,46 +22,60 @@ class ModifyUserType extends AbstractType
         $builder
         ->add('username', TextType::class, [
             'required' => true,
+            'label' => 'Username',
             'attr' =>[
                 'pattern' => '[a-zA-Z0-9]*',
                 'minLength' => 3,
                 'maxLength' => 20,
+                'placeholder' => 'Username',
             ]
         ])
         ->add('firstname', TextType::class, [
             'required' => true,
+            'label' => 'First Name',
             'attr' =>[
-                'pattern' => '[a-zA-Z]*',
+                'pattern' => '[a-zA-Z ]*',
                 'minLength' => 3,
                 'maxLength' => 20,
+                'placeholder' => 'First Name',
             ]
         ])
         ->add('lastname', TextType::class, [
             'required' => true,
+            'label' => 'Last Name',
             'attr' =>[
-                'pattern' => '[a-zA-Z]*',
+                'pattern' => '[a-zA-Z ]*',
                 'minLength' => 3,
                 'maxLength' => 20,
+                'placeholder' => 'Last Name',
             ]
         ])
         ->add('dateNaiss', DateType::class, [
             'widget' => 'single_text',
+            'label' => 'Date of Birth',
             'required' => true,
+            'attr' =>[
+                'placeholder' => 'Date of Birth',
+            ]
         ])
         ->add('numTel', TelType::class, [
             'required' => true,
+            'label' => 'Phone Number',
             'attr' =>[
                 'pattern' => '[0-9]*',
                 'minLength' => 8,
                 'maxLength' => 8,
+                'placeholder' => 'Phone Number',
             ]
         ])
         ->add('adresse', TextType::class, [
             'required' => true,
+            'label' => 'Address',
             'attr' =>[
-                'pattern' => '[a-zA-Z0-9]*',
+                'pattern' => '[a-zA-Z0-9 ]*',
                 'minLength' => 3,
                 'maxLength' => 50,
+                'placeholder' => 'Address',
             ]
         ])
         ->add('submit', SubmitType::class, [
@@ -74,10 +88,7 @@ class ModifyUserType extends AbstractType
     {
         $resolver->setDefaults([
             'data_class' => User::class,
-            'validation_groups' => function (FormInterface $form) {
-                $data = $form->getData();
-                return null === $data->getId() ? ['Default', 'create'] : ['Default', 'update'];
-            },
+            'validation_groups' => ['update' , 'Default'],
         ]);
     }
 }
